fix integer conversion bug in updateAssetsIn (#169)

+++ contracts/Comet.sol
@@ -125,7 +125,9 @@ contract Comet is CometMath, CometStorage {
     uint104 public immutable targetReserves;
 
     /// @notice The max number of assets this contract is hardcoded to support
-    /// @dev Do not change this variable without updating all the fields throughout the contract.
+    /// @dev Do not change this variable without updating all the fields
+    /// throughout the contract (including the size of UserBasic.assetsIn and
+    /// the integer conversion in isInAsset and updateAssetsIn)
     uint8 public constant maxAssets = 15;
 
     /// @notice The number of assets this contract actually supports
@@ -915,7 +917,7 @@ contract Comet is CometMath, CometStorage {
      * @dev Whether user has a non-zero balance of an asset, given assetsIn flags
      */
     function isInAsset(uint16 assetsIn, uint8 assetOffset) internal pure returns (bool) {
-        return (assetsIn & (uint8(1) << assetOffset) != 0);
+        return (assetsIn & (uint16(1) << assetOffset) != 0);
     }
 
     /**
@@ -930,10 +932,10 @@ contract Comet is CometMath, CometStorage {
         AssetInfo memory assetInfo = getAssetInfoByAddress(asset);
         if (initialUserBalance == 0 && finalUserBalance != 0) {
             // set bit for asset
-            userBasic[account].assetsIn |= (uint8(1) << assetInfo.offset);
+            userBasic[account].assetsIn |= (uint16(1) << assetInfo.offset);
         } else if (initialUserBalance != 0 && finalUserBalance == 0) {
             // clear bit for asset
-            userBasic[account].assetsIn &= ~(uint8(1) << assetInfo.offset);
+            userBasic[account].assetsIn &= ~(uint16(1) << assetInfo.offset);
         }
     }
